Level 2: Model for Restaurants + updates/improvements to existing models
This introduces the "restaurant" model. A "restaurant" is the new
top-level object under this model, and can have many menus, which
can all have many items.

Furthermore, menu items are made unique by name and thus can be shared
across menus. This is demonstrated in a restaurant spec that has
separate breakfast and lunch menus, with a unified brunch menu for
weekends.

The whole suite also contains some improvements I missed in the first
round (e.g. automated tests for the price-to-string method in the
menu item model)